pyproject.toml
[project]
name = "tru"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2.7",
    "django-ratelimit==4.1.0",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "drf-spectacular==0.28.0",
    "drf-spectacular-sidecar==2025.10.1",
    "gunicorn==23.0.0",
    "python-dotenv==1.1.1",
    "redis==6.4.0",
    "requests>=2.32.5",
    "setuptools==80.9.0",
    "wheel==0.45.1",
]
